requirements.txt
# 腾河项目：纯 Python 通用依赖（不含 PyTorch）
# PyTorch 请按 README 单独安装：本机 CPU 用 conda/pip cpuonly，云上 GPU 用 CUDA 版。

pandas>=1.4.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
openpyxl>=3.0.0
# P5 特征配置化
PyYAML>=6.0
